Change to explicit .mjs for implementation files
