+++ lib/argument.mjs
@@ -1,4 +1,4 @@
-import { InvalidArgumentError } from './error.js';
+import { InvalidArgumentError } from './error.mjs';
 
 export class Argument {
   /**
