+++ lib/command.mjs
@@ -4,11 +4,11 @@ import path from 'node:path';
 import fs from 'node:fs';
 import process from 'node:process';
 
-import { Argument, humanReadableArgName } from './argument.js';
-import { CommanderError } from './error.js';
-import { Help, stripColor } from './help.js';
-import { Option, DualOptions } from './option.js';
-import { suggestSimilar } from './suggestSimilar.js';
+import { Argument, humanReadableArgName } from './argument.mjs';
+import { CommanderError } from './error.mjs';
+import { Help, stripColor } from './help.mjs';
+import { Option, DualOptions } from './option.mjs';
+import { suggestSimilar } from './suggestSimilar.mjs';
 
 export class Command extends EventEmitter {
   /**
