+++ lib/help.mjs
@@ -1,11 +1,11 @@
-import { humanReadableArgName } from './argument.js';
+import { humanReadableArgName } from './argument.mjs';
 
 /**
  * TypeScript import types for JSDoc, used by Visual Studio Code IntelliSense and `npm run typescript-checkJS`
  * https://www.typescriptlang.org/docs/handbook/jsdoc-supported-types.html#import-types
- * @typedef { import("./argument.js").Argument } Argument
- * @typedef { import("./command.js").Command } Command
- * @typedef { import("./option.js").Option } Option
+ * @typedef { import("./argument.mjs").Argument } Argument
+ * @typedef { import("./command.mjs").Command } Command
+ * @typedef { import("./option.mjs").Option } Option
  */
 
 // Although this is a class, methods are static in style to allow override using subclass or just functions.
